Move all the services to a folder
