+++ photo-sharing-application-front-end/src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { CsrfTokenService } from './csrf-token/csrf-token.service';
-import { AuthService } from './auth/auth.service';
+import { CsrfTokenService } from './service/csrf-token/csrf-token.service';
+import { AuthService } from './service/auth/auth.service';
 
 @Component({
   selector: 'app-root',
