+++ photo-sharing-application-front-end/src/app/service/auth/auth.service.spec.ts
