+++ photo-sharing-application-front-end/src/app/service/auth/auth.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, Signal, signal } from '@angular/core';
-import { Account } from '../model/account';
+import { Account } from '../../model/account';
 import { HttpClient } from '@angular/common/http';
 
 @Injectable({
