+++ photo-sharing-application-front-end/src/app/service/comment/comment.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Page } from '../model/page';
+import { Page } from '../../model/page';
 
 @Injectable({
   providedIn: 'root'
