+++ photo-sharing-application-front-end/src/app/service/csrf-token/csrf-token.service.ts
