+++ photo-sharing-application-front-end/src/app/service/forgotten-password/forgotten-password.spec.ts
