+++ photo-sharing-application-front-end/src/app/service/image/image.service.spec.ts
