+++ photo-sharing-application-front-end/src/app/service/login/login.service.spec.ts
