+++ photo-sharing-application-front-end/src/app/service/registration/registration.service.spec.ts
