+++ photo-sharing-application-front-end/src/app/service/verification/verification.service.spec.ts
